+++ tsconfig.json
@@ -61,8 +61,7 @@
       "@svg/*": ["./src/svg/*"],
       "@hooks/*": ["./src/hooks/*"],
       "@utils": ["./src/utils/index.tsx"],
-      "@providers/*": ["./src/providers/*"],
-      "@web3/*": ["./src/web3/*"]
+      "@providers/*": ["./src/providers/*"]
     },
     // "rootDirs": [],                        /* List of root folders whose combined content represents the structure of the project at runtime. */
     // "typeRoots": [],                       /* List of folders to include type definitions from. */
